# docs/contracts/openapi.v1.json
{
  "components": {
    "schemas": {
      "ComparePayload": {
        "properties": {
          "left": {
            "$ref": "#/components/schemas/FieldPayload"
          },
          "right": {
            "$ref": "#/components/schemas/FieldPayload"
          }
        },
        "required": [
          "left",
          "right"
        ],
        "title": "ComparePayload",
        "type": "object"
      },
      "FieldPayload": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "FieldPayload",
        "type": "object"
      },
      "ForecastPayload": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "horizon": {
            "default": 8,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ForecastPayload",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ReportPayload": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "history": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "items": {
                      "type": "number"
                    },
                    "type": "array"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "horizon": {
            "default": 8,
            "minimum": 1.0,
            "title": "Horizon",
            "type": "integer"
          },
          "output_root": {
            "default": "artifacts/runs",
            "title": "Output Root",
            "type": "string"
          },
          "spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ReportPayload",
        "type": "object"
      },
      "SimulationPayload": {
        "properties": {
          "alpha": {
            "default": 0.18,
            "exclusiveMinimum": 0.0,
            "maximum": 0.25,
            "title": "Alpha",
            "type": "number"
          },
          "grid_size": {
            "default": 32,
            "minimum": 4.0,
            "title": "Grid Size",
            "type": "integer"
          },
          "jitter_var": {
            "default": 0.0005,
            "maximum": 0.01,
            "minimum": 0.0,
            "title": "Jitter Var",
            "type": "number"
          },
          "quantum_jitter": {
            "default": false,
            "title": "Quantum Jitter",
            "type": "boolean"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          },
          "spike_probability": {
            "default": 0.25,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Spike Probability",
            "type": "number"
          },
          "steps": {
            "default": 24,
            "minimum": 1.0,
            "title": "Steps",
            "type": "integer"
          },
          "turing_enabled": {
            "default": true,
            "title": "Turing Enabled",
            "type": "boolean"
          },
          "turing_threshold": {
            "default": 0.75,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Turing Threshold",
            "type": "number"
          },
          "with_history": {
            "default": false,
            "title": "With History",
            "type": "boolean"
          }
        },
        "title": "SimulationPayload",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Morphology-aware Field Intelligence Engine",
    "version": "4.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/compare": {
      "post": {
        "operationId": "compare_endpoint_v1_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Compare Endpoint V1 Compare Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    },
    "/v1/detect": {
      "post": {
        "operationId": "detect_endpoint_v1_detect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Detect Endpoint V1 Detect Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detect Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    },
    "/v1/extract": {
      "post": {
        "operationId": "extract_endpoint_v1_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Extract Endpoint V1 Extract Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    },
    "/v1/forecast": {
      "post": {
        "operationId": "forecast_endpoint_v1_forecast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Forecast Endpoint V1 Forecast Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    },
    "/v1/report": {
      "post": {
        "operationId": "report_endpoint_v1_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Report Endpoint V1 Report Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    },
    "/v1/simulate": {
      "post": {
        "operationId": "simulate_endpoint_v1_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Simulate Endpoint V1 Simulate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Simulate Endpoint",
        "tags": [
          "mfn-v1"
        ]
      }
    }
  }
}